pyproject.toml
[tool.poetry]
name = "track-store-back"
version = "0.1.0"
description = ""
authors = ["Aliev Nurlan <nualiev420@gmail.com>"]
readme = "README.md"
packages = [{include = "goods_map"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
psycopg2-binary = "^2.9.7"
django-cors-headers = "^4.2.0"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
dj-database-url = "^2.1.0"
django-filter = "^23.2"
pillow = "^10.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
